DNS: no separate NS for each domain
Having a separate nameserver for each domain is bad style. Having the nameserver on the very same domain it is governing is very bad style (and not allowed for some TLDs without DNS IP fixing).

Instead, the main server (of the EasySCP installation) gets to be the the main nameserver for all domains. Not the perfect solution either, I will expand on it in the PR.